COSMIT Make it easier to add config settings
